skill_taxonomy_pipeline/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Embeddings and NLP
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
nltk>=3.8.0
spacy>=3.5.0

# Efficient similarity search
faiss-cpu>=1.7.4
annoy>=1.17.0

# Clustering algorithms
hdbscan>=0.8.33
umap-learn>=0.5.4

# LLM Integration
openai>=1.0.0
tiktoken>=0.5.0
langchain>=0.1.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
networkx>=3.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rapidfuzz>=3.0.0
joblib>=1.3.0

# Data validation
pandera>=0.15.0
great-expectations>=0.17.0
